Changed file paths

+++ src/utility/mathutility.cpp
@@ -1,4 +1,4 @@
-#include <mathutility.h>
+#include <utility/mathutility.h>
 
 float MathUtility::RandomFloat() {
     return (float)rand() / RAND_MAX;
@@ -6,5 +6,5 @@ float MathUtility::RandomFloat() {
 
 float MathUtility::Lerp(float a, float b, float step)
 {
-    return a * (1.0 - step) + (b * step);
+    return (float)(a * (1.0 - step) + (b * step));
 }
